Added more test cases
Finished the test cases for angle conversions.
Started adding test cases for approximate equality utility (integers
done so far -- still to do floating point).
Renamed "circles" to "turns" in angle conversion so that it's clearer.
Fixed a couple of utility functions which wouldn't compile as constexpr
on VS2015 due to multiple return statements.