Minor refactor of BoundingBox2d
Changed the name of a member and some function parameters in
BoundingBox2d to make them clearer, and added an initializer list
constructor. (Note that I haven't actually compiled this class into any
test code yet so it may be broken at the moment.)
Made a minor change to the wording of exceptions on the initializer list
constructors in Vector2d and Polar.